move:#27 user-meta-data 폴더 분리 - my-page 내부 depth가 깊어져 feature 하단으로 분리
